requirements.txt
requests>=2.31.0
qrcode>=7.4.2
pillow>=10.0.0
rumps>=0.4.0
